chore: rename the PanelElements file to PanelElement
