+++ packages/website/src/components/panel/accordionContent/EdgeDetection.tsx
@@ -1,5 +1,5 @@
 import { AccordionContent } from '../../ui/accordion';
-import PanelElement from '../PanelElements';
+import PanelElement from '../PanelElement';
 import { Controller } from '../panel';
 
 export default function EdgeDetection() {
