+++ packages/website/src/components/panel/accordionContent/svgControls/RECURSIVE.tsx
@@ -1,5 +1,5 @@
 import { Controller } from '../../panel';
-import PanelElement from '../../PanelElements';
+import PanelElement from '../../PanelElement';
 
 export default function RECURSIVE() {
 	const imageControls: Controller[] = [
